% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcost.R
\name{getCost}
\alias{getCost}
\title{Return cost of trade}
\usage{
getCost(volume, spread)
}
\arguments{
\item{volume}{Volume of trade in millions}

\item{spread}{Spread of trade in percentage}
}
\value{
Cost in millions
}
\description{
Return cost of trade
}
