% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcost.R
\name{getSpread}
\alias{getSpread}
\title{Return spread for a given bond trade}
\usage{
getSpread(term, volume, instrument, liquidity)
}
\arguments{
\item{term}{Bond's term to maturity in years}

\item{volume}{Volume of trade in millions}

\item{instrument}{Name of instrument}

\item{liquidity}{Liquidity data object}
}
\value{
Spread in percentage
}
\description{
Return spread for a given bond trade
}
