% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcost.R
\name{loadLiquidity}
\alias{loadLiquidity}
\title{Load liquidity data from files}
\usage{
loadLiquidity(file = c("liq_data/liq_t.csv", "liq_data/liq_b.csv"),
  instrument = c("T", "B"))
}
\arguments{
\item{file}{Names of csv data files}

\item{instrument}{Names of instruments}
}
\value{
Liquidity data object
}
\description{
Attributes:
spreads - spreads in percentage;
terms - terms of bonds;
volumes - trade volumes
}
